fixed responsive design on 'How it works' block
