--- src/containers/user-how-it-works/TutorHowItWorksCards.js
@@ -1,31 +1,0 @@
-import howItWorksTutorFirst from '~/assets/img/guest-home-page/howItWorksTutorFirst.svg'
-import howItWorksTutorSecond from '~/assets/img/guest-home-page/howItWorksTutorSecond.svg'
-import howItWorksTutorThird from '~/assets/img/guest-home-page/howItWorksTutorThird.svg'
-import howItWorksTutorFourth from '~/assets/img/guest-home-page/howItWorksTutorFourth.svg'
-
-export const tutorHowItWorksCards = [
-  {
-    image: howItWorksTutorFirst,
-    title: 'tutorHomePage.howItWorks.signUp.title',
-    description: 'tutorHomePage.howItWorks.signUp.description',
-    id: 1
-  },
-  {
-    image: howItWorksTutorSecond,
-    title: 'tutorHomePage.howItWorks.createATutorAccount.title',
-    description: 'tutorHomePage.howItWorks.createATutorAccount.description',
-    id: 2
-  },
-  {
-    image: howItWorksTutorThird,
-    title: 'tutorHomePage.howItWorks.getNewStudents.title',
-    description: 'tutorHomePage.howItWorks.getNewStudents.description',
-    id: 3
-  },
-  {
-    image: howItWorksTutorFourth,
-    title: 'tutorHomePage.howItWorks.receiveFeedbacks.title',
-    description: 'tutorHomePage.howItWorks.receiveFeedbacks.description',
-    id: 4
-  }
-]
