+++ src/containers/user-how-it-works/UserHowItWorksCards.js
@@ -1,3 +1,7 @@
+import howItWorksTutorFirst from '~/assets/img/guest-home-page/howItWorksTutorFirst.svg'
+import howItWorksTutorSecond from '~/assets/img/guest-home-page/howItWorksTutorSecond.svg'
+import howItWorksTutorThird from '~/assets/img/guest-home-page/howItWorksTutorThird.svg'
+import howItWorksTutorFourth from '~/assets/img/guest-home-page/howItWorksTutorFourth.svg'
 import howItWorksStudentFirst from '~/assets/img/guest-home-page/howItWorksStudentFirst.svg'
 import howItWorksStudentSecond from '~/assets/img/guest-home-page/howItWorksStudentSecond.svg'
 import howItWorksStudentThird from '~/assets/img/guest-home-page/howItWorksStudentThird.svg'
@@ -26,6 +30,33 @@ export const studentHowItWorksCards = [
     image: howItWorksStudentFirst,
     title: 'studentHomePage.howItWorks.writeFeedback.title',
     description: 'studentHomePage.howItWorks.writeFeedback.description',
+    id: 4
+  }
+]
+
+export const tutorHowItWorksCards = [
+  {
+    image: howItWorksTutorFirst,
+    title: 'tutorHomePage.howItWorks.signUp.title',
+    description: 'tutorHomePage.howItWorks.signUp.description',
+    id: 1
+  },
+  {
+    image: howItWorksTutorSecond,
+    title: 'tutorHomePage.howItWorks.createATutorAccount.title',
+    description: 'tutorHomePage.howItWorks.createATutorAccount.description',
+    id: 2
+  },
+  {
+    image: howItWorksTutorThird,
+    title: 'tutorHomePage.howItWorks.getNewStudents.title',
+    description: 'tutorHomePage.howItWorks.getNewStudents.description',
+    id: 3
+  },
+  {
+    image: howItWorksTutorFourth,
+    title: 'tutorHomePage.howItWorks.receiveFeedbacks.title',
+    description: 'tutorHomePage.howItWorks.receiveFeedbacks.description',
     id: 4
   }
 ]
